requirements.txt
Flask (0.10.1)
Flask-WTF (0.12)
gunicorn (19.4.5)
itsdangerous (0.24)
Jinja2 (2.8)
MarkupSafe (0.23)
passlib (1.6.5)
pip (1.5.4)
PyJWT (1.4.0)
pymongo (3.2.1)
requests (2.9.1)
setuptools (2.2)
Werkzeug (0.11.4)
wsgiref (0.1.2)
WTForms (2.1)
